+++ src/OpenTelemetry.OpAmp.Client/CHANGELOG.md
@@ -18,6 +18,9 @@
 * Add settings for remote configuration and update advertised capabilities.
   ([#3618](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3618))
 
+* Expose public `CustomCapabilitiesMessage` and `CustomMessageMessage`.
+  ([#TODO](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/TODO))
+
 ## 0.1.0-alpha.3
 
 Released 2025-Nov-13